+++ pkg/analysis_server/lib/src/lsp/handlers/handler_document_symbols.dart
@@ -40,7 +40,7 @@ class DocumentSymbolHandler extends SharedMessageHandler<DocumentSymbolParams,
     final path = pathOfDoc(params.textDocument);
     final unit = await path.mapResult(requireResolvedUnit);
     return unit.mapResultSync(
-        (unit) => _getSymbols(clientCapabilities, path.result, unit));
+        (unit) => _getSymbols(clientCapabilities, unit.path, unit));
   }
 
   DocumentSymbol _asDocumentSymbol(