+++ pkg/meta/lib/meta_meta.dart
@@ -53,11 +53,11 @@ class TargetKind {
   /// `part of` directive.
   static const directive = TargetKind._('directives', 'directive');
 
-  /// Indicates that an annotation is valid on any enum value declaration.
-  static const enumValue = TargetKind._('enum values', 'enumValue');
-
   /// Indicates that an annotation is valid on any enum declaration.
   static const enumType = TargetKind._('enums', 'enumType');
+
+  /// Indicates that an annotation is valid on any enum value declaration.
+  static const enumValue = TargetKind._('enum values', 'enumValue');
 
   /// Indicates that an annotation is valid on any extension declaration.
   static const extension = TargetKind._('extensions', 'extension');
@@ -128,8 +128,8 @@ class TargetKind {
     classType,
     constructor,
     directive,
-    enumValue,
     enumType,
+    enumValue,
     extension,
     extensionType,
     field,