+++ pkg/dds/CHANGELOG.md
@@ -3,6 +3,8 @@
 and prepared DDS for using `unified_analytics` through the Dart Tooling Daemon.
 - Internal change: removed `analytics` parameter from the DevTools server `defaultHandler` method.
 - Updated `README.md` and added  contributing guide (`CONTRIBUTING.md`).
+- Updated `package:dds_service_extensions` constraint to ^2.0.0.
+- Determine default `requireUserPermissionToResume` values from the `pause_isolates_on_start` and `pause_isolates_on_exit` flags.
 
 # 4.0.0
 - Updated DDS protocol to version 2.0.